+++ Core/Systems/VanillaModifications/BossItemRemovals/BossGlobalItemDisabler.cs
@@ -1,12 +1,18 @@
 ﻿using System.Collections.Generic;
 using Terraria.ID;
 
-namespace PathOfTerraria.Core.Systems.VanillaModifications;
+namespace PathOfTerraria.Core.Systems.VanillaModifications.BossItemRemovals;
 
 internal class BossGlobalItemDisabler : GlobalItem
 {
-	public static readonly HashSet<int> BossSpawners = [ItemID.SlimeCrown, ItemID.SuspiciousLookingEye, ItemID.WormFood, ItemID.BloodySpine, ItemID.Abeemination, 
-		ItemID.ClothierVoodooDoll, ItemID.GuideVoodooDoll, ItemID.DeerThing]; 
+	public static readonly HashSet<int> BossSpawners = [ItemID.SlimeCrown,
+		ItemID.SuspiciousLookingEye,
+		ItemID.WormFood,
+		ItemID.BloodySpine,
+		ItemID.Abeemination,
+		ItemID.ClothierVoodooDoll,
+		ItemID.GuideVoodooDoll,
+		ItemID.DeerThing];
 
 	public override bool AppliesToEntity(Item entity, bool lateInstantiation)
 	{
